Simplify roundtrip test
