+++ identity_jose/src/tests/mod.rs
@@ -9,3 +9,4 @@ mod rfc7517;
 mod rfc7638;
 mod rfc7797;
 mod rfc8037;
+mod roundtrip;
